Move a bunch of data structures from farmer binary to library and make them public
