+++ crates/subspace-farmer/src/bin/subspace-farmer/main.rs
@@ -1,4 +1,4 @@
-#![feature(const_option, type_alias_impl_trait, type_changing_struct_update)]
+#![feature(const_option, type_changing_struct_update)]
 
 mod commands;
 mod ss58;
