+++ crates/subspace-farmer/src/utils.rs
@@ -1,5 +1,11 @@
+pub mod farmer_piece_cache;
+pub mod farmer_piece_getter;
+pub mod farmer_provider_storage;
+pub mod node_piece_getter;
 pub mod parity_db_store;
+pub mod piece_cache;
 pub mod piece_validator;
+pub mod readers_and_pieces;
 
 use std::ops::Deref;
